fix: changed the name to EntityMigrationFramework
