+++ .github/workflows/publish.yml
@@ -1,0 +1,41 @@
+name: Publish to NuGet
+
+on:
+  workflow_run:
+    workflows: 
+      - "Run Unit Tests"
+      - "Check Code Formatting"
+    types: [completed]
+  push:
+    branches:
+      - master
+    tags:
+      - 'v*'
+    
+
+jobs:
+  deploy:
+    if: github.event.workflow_run.conclusion == 'success'
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    
+    - name: Setup .NET
+      uses: actions/setup-dotnet@v2
+      with:
+        dotnet-version: '8.0.x'
+    
+    - name: Restore
+      run: dotnet restore
+    
+    - name: Build
+      run: dotnet build --configuration Release --no-restore
+    
+    - name: Test
+      run: dotnet test --no-restore --verbosity normal
+    
+    - name: Pack
+      run: dotnet pack --configuration Release --no-build -o out
+    
+    - name: Push to NuGet
+      run: dotnet nuget push ./out/*.nupkg --api-key ${{secrets.NUGET_API_KEY}} --source https://api.nuget.org/v3/index.json --skip-duplicate
