--- EntityMigration.Tests/UnitTest1.cs
@@ -1,7 +1,0 @@
-﻿namespace EntityMigration.Tests;
-
-public class UnitTest1
-{
-    [Fact]
-    public void Test1() { }
-}
