--- EntityMigrationExampleApp/Model/User/Versions/UserV1.cs
@@ -1,12 +1,0 @@
-namespace EntityMigrationExampleApp.Model.User.Versions;
-
-/// <summary>
-/// Initial user model with basic name property
-/// </summary>
-public class UserV1 : BaseUser
-{
-    /// <summary>
-    /// Full name of the user
-    /// </summary>
-    public required string Name { get; set; }
-}
