+++ EntityMigrationFramework.Tests/UnitTest1.cs
@@ -1,0 +1,7 @@
+﻿namespace EntityMigrationFramework.Tests;
+
+public class UnitTest1
+{
+    [Fact]
+    public void Test1() { }
+}
